+++ robot/src/test/java/us/ilite/robot/driverinput/DriverInputTest.java
@@ -7,10 +7,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.*;
 import org.mockito.junit.MockitoJUnitRunner;
+import us.ilite.common.Data;
 import us.ilite.common.config.SystemSettings;
 import us.ilite.common.types.input.ELogitech310;
 import us.ilite.lib.drivers.Clock;
-import us.ilite.robot.RobotData;
 import us.ilite.robot.commands.Delay;
 import us.ilite.robot.modules.Drive;
 import us.ilite.robot.modules.ModuleList;
@@ -30,7 +30,7 @@ public class DriverInputTest {
 
     private DriverInput mDriverInput;
 
-    private RobotData mData;
+    private Data mData;
     private Clock mClock;
     private ModuleList mModuleList;
 
@@ -39,7 +39,7 @@ public class DriverInputTest {
         MockitoAnnotations.initMocks(this);
         Logger.setLevel(ELevel.DEBUG);
 
-        mData = new RobotData();
+        mData = new Data();
         mClock = new Clock().simulated();
         mModuleList = new ModuleList();
         mSuperstructure = spy(new Superstructure());
